+++ libs/core/src/integration/events.schemas.ts
@@ -262,3 +262,7 @@ export const FarcasterReplyCastCreated = FarcasterCast.describe(
 export const FarcasterVoteCreated = FarcasterAction.extend({
   contest_address: z.string(),
 }).describe('When a farcaster action is initiated on a cast reply');
+
+export const CheckContests = z.object({});
+
+export const RolloverContests = z.object({});
