+++ libs/model/src/policies/ContestWorker.policy.ts
@@ -1,6 +1,11 @@
 import { Actor, events, logger, Policy } from '@hicommonwealth/core';
+import { NeynarAPIClient } from '@neynar/nodejs-sdk';
+import moment from 'moment';
 import { QueryTypes } from 'sequelize';
-import { Contest, models } from '..';
+import Web3 from 'web3';
+import { config, Contest, models } from '..';
+import { GetActiveContestManagers } from '../contest';
+import { rollOverContest } from '../services/commonProtocol/contestHelper';
 import { buildThreadContentUrl, getChainNodeUrl } from '../utils';
 import {
   createOnchainContestContent,
@@ -12,6 +17,16 @@ const log = logger(import.meta);
 const inputs = {
   ThreadCreated: events.ThreadCreated,
   ThreadUpvoted: events.ThreadUpvoted,
+  CheckContests: events.CheckContests,
+  RolloverContests: events.RolloverContests,
+};
+
+const getPrivateWalletAddress = () => {
+  const web3 = new Web3();
+  const privateKey = config.WEB3.PRIVATE_KEY;
+  const account = web3.eth.accounts.privateKeyToAccount(privateKey);
+  const publicAddress = account.address;
+  return publicAddress;
 };
 
 export function ContestWorker(): Policy<typeof inputs> {
@@ -128,6 +143,171 @@ export function ContestWorker(): Policy<typeof inputs> {
           content_url,
           author_address: payload.address!,
         });
+      },
+      CheckContests: async ({ payload }) => {
+        const activeContestManagers = await GetActiveContestManagers().body({
+          actor: {} as Actor,
+          payload: {},
+        });
+        // find active contests that have content with no upvotes and will end in one hour
+        const contestsWithoutVote = activeContestManagers!.filter(
+          (contestManager) =>
+            contestManager.actions.some(
+              (action) => action.action === 'added',
+            ) &&
+            !contestManager.actions.some(
+              (action) => action.action === 'upvoted',
+            ) &&
+            moment(contestManager.end_time).diff(moment(), 'minutes') < 60,
+        );
+
+        const promises = contestsWithoutVote.map(async (contestManager) => {
+          // add onchain vote to the first content
+          const firstContent = contestManager.actions.find(
+            (action) => action.action === 'added',
+          );
+
+          await createOnchainContestVote({
+            contestManagers: [
+              {
+                url: contestManager.url,
+                contest_address: contestManager.contest_address,
+                content_id: firstContent!.content_id,
+              },
+            ],
+            content_url: firstContent!.content_url!,
+            author_address: getPrivateWalletAddress(),
+          });
+        });
+
+        const promiseResults = await Promise.allSettled(promises);
+
+        const errors = promiseResults
+          .filter(({ status }) => status === 'rejected')
+          .map(
+            (result) =>
+              (result as PromiseRejectedResult).reason || '<unknown reason>',
+          );
+
+        if (errors.length > 0) {
+          log.error(`CheckContests: failed with errors: ${errors.join(', ')}"`);
+        }
+      },
+      RolloverContests: async ({ payload }) => {
+        const contestManagersWithEndedContest = await models.sequelize.query<{
+          contest_address: string;
+          interval: number;
+          ended: boolean;
+          url: string;
+          private_url: string;
+          neynar_webhook_id?: string;
+        }>(
+          `
+            SELECT cm.contest_address,
+                   cm.interval,
+                   cm.ended,
+                   cm.neynar_webhook_id,
+                   co.end_time,
+                   cn.private_url,
+                   cn.url
+            FROM "ContestManagers" cm
+                     JOIN (SELECT *
+                           FROM "Contests"
+                           WHERE (contest_address, contest_id) IN (SELECT contest_address, MAX(contest_id) AS contest_id
+                                                                   FROM "Contests"
+                                                                   GROUP BY contest_address)) co
+                          ON co.contest_address = cm.contest_address
+                              AND (
+                                 (cm.interval = 0 AND cm.ended IS NOT TRUE)
+                                     OR
+                                 cm.interval > 0
+                                 )
+                              AND NOW() > co.end_time
+                              AND cm.cancelled IS NOT TRUE
+                     JOIN "Communities" cu ON cm.community_id = cu.id
+                     JOIN "ChainNodes" cn ON cu.chain_node_id = cn.id;
+        `,
+          {
+            type: QueryTypes.SELECT,
+            raw: true,
+          },
+        );
+
+        const contestRolloverPromises = contestManagersWithEndedContest.map(
+          async ({
+            url,
+            private_url,
+            contest_address,
+            interval,
+            ended,
+            neynar_webhook_id,
+          }) => {
+            log.info(`ROLLOVER: ${contest_address}`);
+
+            if (interval === 0 && !ended) {
+              // preemptively mark as ended so that rollover
+              // is not attempted again after failure
+              await models.ContestManager.update(
+                {
+                  ended: true,
+                },
+                {
+                  where: {
+                    contest_address,
+                  },
+                },
+              );
+            }
+
+            await rollOverContest(
+              getChainNodeUrl({ url, private_url }),
+              contest_address,
+              interval === 0,
+            );
+
+            // clean up neynar webhooks when farcaster contest ends
+            if (neynar_webhook_id) {
+              try {
+                const client = new NeynarAPIClient(
+                  config.CONTESTS.NEYNAR_API_KEY!,
+                );
+                await client.deleteWebhook(neynar_webhook_id);
+                await models.ContestManager.update(
+                  {
+                    neynar_webhook_id: null,
+                    neynar_webhook_secret: null,
+                  },
+                  {
+                    where: {
+                      contest_address,
+                    },
+                  },
+                );
+              } catch (err) {
+                log.warn(
+                  `failed to delete neynar webhook: ${neynar_webhook_id}`,
+                );
+              }
+            }
+          },
+        );
+
+        const promiseResults = await Promise.allSettled(
+          contestRolloverPromises,
+        );
+
+        const errors = promiseResults
+          .filter(({ status }) => status === 'rejected')
+          .map(
+            (result) =>
+              (result as PromiseRejectedResult).reason || '<unknown reason>',
+          );
+
+        if (errors.length > 0) {
+          log.error(
+            `PerformContestRollovers: failed with errors: ${errors.join(', ')}"`,
+          );
+        }
       },
     },
   };
